+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Update graph colors
+* New stat: total participants by status
 
 ### 1.8.2: October 11th, 2021
 * New stat: Past Bookings by date and by type